Changed login validation.
Working on making the login validation work like we want. There is a problem with verifying the credentials correctly. At the moment all of the hashed passwords in SQuirreL were hashed via MySQL. I have a feeling the way MySQL is hashing the password is different than how it is done in our C# backend. Once we have our account creation method created in the C# backend, and we use the same hashing method used in creating the account when validating the password on login, I think the login validation will work correctly.