Changed how item popup closes.
I made it so that the popups that show up in the drinkmenu and foodmenu
when you click on an item close when you hit the back button, and an
alert is shown when you click the "Add to Order" button (along with closing
the popup). I am going to add a small 'X' button at the top of the popup
to close the popup by clicking on it. I tried making it so that the popup
closes when you click outside of it, but I was having trouble.