Fixed menu searching error.
Previously there was an error when searching the menus for an item. When you looked up an item, it would get rid of items that don't match that searched up item. So when you cancel your current search, you were still left with the results of your previous search. This has been fixed by having a master list of all the data that is never modified and a currentDataDisplayed list to specify which items are currently being displayed on the menu based on the search bar.